requirements.txt
pandas>=2.0
pyyaml
python-dotenv